Split tcp-proxy into separate lib
